chore: extract Expecter into its own package
