+++ testutil/expecter/stdbuf_internal_test.go
@@ -1,4 +1,4 @@
-package ptytest
+package expecter
 
 import (
 	"bytes"
